refactor files to the core module
